backend/pyproject.toml
[project]
name = "team-performance-hub-backend"
version = "0.1.0"
description = "Дашборд активности разработчиков на основе GitLab"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "python-dotenv==1.0.0",
    "aiohttp==3.9.0",
    "gidgetlab==2.1.1",
    "openai==1.3.0",
    "pydantic==2.5.0",
    "cachetools>=7.0.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
